httpclient

+++ TradeTrack/Models/EmployeeInput.swift
@@ -3,6 +3,6 @@ import Foundation
 struct EmployeeInput: Codable {
     let employeeId: String
     let name: String
-    let embedding: [Double]
+    let embedding: [Float]
     let role: String
 }
